Refactor command logic into separate file
game.c should only contain functions related to handling the
data structure. In fact, there are several functions on the data
structure that should be refactored so fields of game_t don't
need to be exposed to cmd.c. That might be too much of a headache
though so I will merge this for now